Fix mixing server stuff and client stuff, resulting in errors
